devcontainer feature

+++ src/base/scripts/aws.sh
@@ -23,11 +23,16 @@ apt install -y --no-install-recommends gh
 
 # Install exercism-cli
 EXERCISM_VERSION="${VERSION:-"latest"}"
-find_version_from_git_tags EXERCISM_VERSION https://github.com/exercism/cli
-exercism_filename="exercism-${EXERCISM_VERSION}-linux-x86_64.tar.gz"
-curl -L https://github.com/exercism/cli/releases/download/v${EXERCISM_VERSION}/${exercism_filename} --create-dirs -o /tmp/${exercism_filename}
-tar -xzvf /tmp/${exercism_filename} -C /usr/local/bin exercism
-rm -rf /tmp/${exercism_filename}
+
+if command -v exercism &> /dev/null; then
+    echo "exercism is installed. Version: $(exercism version)"
+else
+  find_version_from_git_tags EXERCISM_VERSION https://github.com/exercism/cli
+  exercism_filename="exercism-${EXERCISM_VERSION}-linux-x86_64.tar.gz"
+  curl -L https://github.com/exercism/cli/releases/download/v${EXERCISM_VERSION}/${exercism_filename} --create-dirs -o /tmp/${exercism_filename}
+  tar -xzvf /tmp/${exercism_filename} -C /usr/local/bin exercism
+  rm -rf /tmp/${exercism_filename}
+fi
 
 # Install AWS
 if command -v aws &> /dev/null; then
@@ -48,17 +53,22 @@ else
 fi
 
 # Install cfn-lint
-#su ${USERNAME} -c "source /etc/zsh/zshrc && pip install cfn-lint"
+su ${USERNAME} -c 'command -v pyenv >/dev/null 2>&1 && pyenv exec pip install -U cfn-lint || echo "pyenv not found. Please install pyenv to use this script."'
 
 # Install terraform
 TERRAFORM_VERSION="${VERSION:-"latest"}"
-# Verify requested version is available, convert latest
-find_version_from_git_tags TERRAFORM_VERSION 'https://github.com/hashicorp/terraform'
 
-terraform_filename="terraform_${TERRAFORM_VERSION}_linux_amd64.zip"
-curl -sSL "https://releases.hashicorp.com/terraform/${TERRAFORM_VERSION}/${terraform_filename}" -o /tmp/${terraform_filename} && cd $(dirname $_)
-unzip ${terraform_filename}
-mv -f terraform /usr/local/bin/
-rm -rf ${terraform_filename}
+if command -v terraform &> /dev/null; then
+    echo "terraform is installed. Version: $(terraform --version)"
+else
+  # Verify requested version is available, convert latest
+  find_version_from_git_tags TERRAFORM_VERSION 'https://github.com/hashicorp/terraform'
+
+  terraform_filename="terraform_${TERRAFORM_VERSION}_linux_amd64.zip"
+  curl -sSL "https://releases.hashicorp.com/terraform/${TERRAFORM_VERSION}/${terraform_filename}" -o /tmp/${terraform_filename} && cd $(dirname $_)
+  unzip ${terraform_filename}
+  mv -f terraform /usr/local/bin/
+  rm -rf ${terraform_filename}
+fi
 
 echo "Done!"
